# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS System API Documentation",
        "description": "Comprehensive Point of Sale System API for multi-tenant restaurant management with authentication, orders, staff, and more.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get list of menu items",
                "description": "Retrieve paginated list of menu items for the authenticated restaurant",
                "operationId": "69254261965b456d128d15f2b7c4cda6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Alias for per_page - Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "is_available",
                        "in": "query",
                        "description": "Filter by availability status (true/false or 1/0)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "true"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search items by name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Caesar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of menu items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MenuItem"
                                            }
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 35
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Create new menu item",
                "description": "Create a new menu item for the authenticated restaurant",
                "operationId": "354fedf84d71e303495b6ca8dd023055",
                "requestBody": {
                    "description": "Menu item data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Caesar Salad"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fresh romaine lettuce with caesar dressing"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "preparation_time": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "calories": {
                                        "type": "integer",
                                        "example": 250
                                    },
                                    "allergens": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "gluten",
                                            "dairy"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get specific menu item",
                "description": "Retrieve details of a specific menu item with category information",
                "operationId": "f0822ac4ed25c71d1ea734ebf1b86dc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu item details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update menu item",
                "description": "Update an existing menu item",
                "operationId": "37145c8f34178f66cdd3ffee4985e34b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated menu item data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Caesar Salad"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 14.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "preparation_time": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Delete menu item",
                "description": "Delete an existing menu item",
                "operationId": "a18f583baa7e079180e185b389677f6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Menu item deleted successfully"
                    },
                    "404": {
                        "description": "Menu item not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MenuCategory": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Appetizers"
                    },
                    "description": {
                        "type": "string",
                        "example": "Light meals to start"
                    },
                    "display_order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "restaurant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MenuItem": {
                "required": [
                    "name",
                    "price",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Caesar Salad"
                    },
                    "description": {
                        "type": "string",
                        "example": "Fresh romaine lettuce with caesar dressing"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 12.99
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "preparation_time": {
                        "type": "integer",
                        "example": 15
                    },
                    "calories": {
                        "type": "integer",
                        "example": 250
                    },
                    "allergens": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "gluten",
                            "dairy"
                        ]
                    },
                    "restaurant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/MenuCategory"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "JWT Bearer token authentication. Add 'Bearer ' before your token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}